chore(search): break up search component
