+++ components/views/chat/search/input/Input.html
@@ -1,5 +1,5 @@
 <div class="search-container" v-click-outside="lostFocus">
-  <div :class="{'search-box': true, focus: isFocus}" class="disabled">
+  <div :class="{'search-box': true, focus: isFocus}">
     <div :class="{'search-placeholder': true, hide: !isEmpty}">
       {{placeholder}}
     </div>
@@ -12,13 +12,13 @@
     <div
       ref="searchInput"
       class="search-input"
-      @mousedown="/* setFocus disable temporarily until search rework */"
+      @mousedown="setFocus"
       @keydown="keydown"
       @keyup="keyup"
       @input="input"
     ></div>
   </div>
-  <div :class="{'search-options': true}" v-if="isOption">
+  <div :class="{'search-options': true}" v-if="false">
     <ul class="search-option-group">
       <li class="search-option-heading">
         {{$t('search.input.search_options')}}
@@ -48,7 +48,7 @@
       <li class="search-for">
         {{$t('search.input.search_for')}} {{searchFor}}
       </li>
-      <li
+      <!-- <li
         v-for="(sch, i) in searchResult"
         :class="{active:i==search}"
         @click="clickSearchOption(i)"
@@ -56,7 +56,7 @@
         <span>{{sch.command}}:</span>
         <span>{{sch.value.value.substring(0, current.value.length)}}</span
         ><span>{{sch.value.value.substring(current.value.length)}}</span>
-      </li>
+      </li> -->
     </ul>
   </div>
   <div
