+++ components/views/chat/search/input/InputGlobal.less
