+++ components/views/chat/search/result/Result.html
@@ -7,16 +7,12 @@
     {{$t('search.result.search_results')}}
     <InteractablesClose class="close" :action="toggle" />
   </div>
-  <div class="search-result-group">
-    <div
-      v-for="item in groupList"
-      :class="{'group-item': true, 'active': item.type === groupBy}"
-      @click="toggleGroupBy(item.type)"
-    >
-      {{item.title}} {{item.count}}
-    </div>
-  </div>
-  <div class="search-result-filters">
+  <SearchResultGroup
+    :groupList="groupList"
+    :groupBy="groupBy"
+    @toggleGroup="toggleGroupBy"
+  />
+  <!-- <div class="search-result-filters">
     <div class="container-filter user-options">
       <multiselect
         :options="userOptions"
@@ -69,24 +65,14 @@
         {{$t('search.result.more_filters')}}
       </button>
     </div>
-  </div>
-  <div
-    class="search-result-meta"
+  </div> -->
+  <SearchResultMeta
     v-if="loading !== DataStateType.Loading && result && result.data && result.data.totalRows"
-  >
-    <div class="result-count">
-      {{result.data.totalRows}} {{$t('search.result.results')}}
-    </div>
-    <div class="result-orderby">
-      <div
-        v-for="item in orderTypeList"
-        :class="{'orderby-item': true, 'active': item.type === orderBy}"
-        @click="toggleOrderBy(item.type)"
-      >
-        {{item.title}}
-      </div>
-    </div>
-  </div>
+    :result="result"
+    :items="orderTypeList"
+    :orderBy="orderBy"
+    @toggleOrder="toggleOrderBy"
+  />
   <UiScroll
     verticalScroll
     scrollbarVisibility="scroll"
@@ -94,13 +80,12 @@
   >
     <div class="search-result-inner">
       <UiLoadersMessage :count="2" v-if="loading === DataStateType.Loading" />
-      <template v-else>
-        <InteractablesSearchResultItem
-          v-for="item, i in result.data.list"
-          :data="item"
-          :key="i"
-        ></InteractablesSearchResultItem>
-      </template>
+      <SearchResultItem
+        v-else
+        v-for="item, i in result.data.list"
+        :data="item"
+        :key="i"
+      />
     </div>
   </UiScroll>
   <paginate
@@ -112,11 +97,5 @@
     container-class="search-result-pagination"
     :click-handler="handleClickPaginate"
   />
-  <div class="search-result-history">
-    <InteractablesChip
-      v-for="(item, i) in givenQueryItems"
-      :key="i"
-      :text="item"
-    />
-  </div>
+  <SearchResultHistory :items="givenQueryItems" />
 </div>
