+++ components/views/chat/search/result/ResultGlobal.less
