+++ components/views/chat/search/result/history/History.html
@@ -1,0 +1,3 @@
+<div class="history">
+  <InteractablesChip v-for="(item, i) in items" :key="i" :text="item" />
+</div>
