+++ components/views/chat/search/result/history/History.less
@@ -1,0 +1,3 @@
+.history {
+  padding: 5px 10px;
+}
