+++ components/views/chat/search/result/item/Item.less
@@ -1,0 +1,26 @@
+.result-item {
+  padding: 10px;
+  margin: 8px 0;
+  &:extend(.round-corners);
+  &:extend(.background-semitransparent-light);
+  box-shadow: @light-blur;
+  border: @light-border !important;
+  display: flex;
+  &:extend(.full-width);
+  .avatar {
+    margin-right: 10px;
+  }
+  .content {
+    max-width: calc(@full - 48px);
+    .username {
+      font-size: @title-text;
+      font-weight: bold;
+    }
+    .message {
+      font-size: @text-size;
+      background: transparent;
+      text-overflow: inherit;
+      white-space: pre-wrap;
+    }
+  }
+}
