+++ components/views/chat/search/result/item/Item.vue
@@ -1,0 +1,19 @@
+<template src="./Item.html"></template>
+
+<script lang="ts">
+import Vue, { PropType } from 'vue'
+import { SearchResultItem } from '~/types/search/search'
+
+declare module 'vue/types/vue' {
+  interface Vue {}
+}
+export default Vue.extend({
+  props: {
+    data: {
+      type: Object as PropType<SearchResultItem>,
+      default: {} as SearchResultItem,
+    },
+  },
+})
+</script>
+<style scoped lang="less" src="./Item.less" />
