+++ components/views/chat/search/result/meta/Meta.html
@@ -1,0 +1,14 @@
+<div class="meta">
+  <div class="result-count">
+    {{result.data.totalRows}} {{$t('search.result.results')}}
+  </div>
+  <div class="result-orderby">
+    <div
+      v-for="item in items"
+      :class="{'orderby-item': true, 'active': item.type === orderBy}"
+      @click="$emit('toggleOrder',item.type)"
+    >
+      {{item.title}}
+    </div>
+  </div>
+</div>
