+++ components/views/chat/search/result/meta/Meta.less
@@ -1,0 +1,21 @@
+.meta {
+  .result-count {
+    font-size: @text-size;
+    font-weight: bold;
+  }
+  .result-orderby {
+    .orderby-item {
+      &.active {
+        font-weight: bold;
+      }
+      cursor: pointer;
+      padding: 2px 8px;
+    }
+    display: flex;
+  }
+  padding: 5px 10px 10px;
+  display: flex;
+  justify-content: space-between;
+  align-items: center;
+  margin: 0px 12px;
+}
