+++ layouts/Layout.less
@@ -148,6 +148,9 @@
     width: calc(@full - (@normal-spacing * 2));
     margin: 0 1rem;
   }
+  .search-result {
+    margin-right: calc(@sidebar-size + @servers-size);
+  }
 }
 
 &.is-open-aside {
